llm-secret-guard/requirements.txt
python-dotenv>=1.0.1
openai>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0
anthropic>=0.7.0
